Cavendish/requirements.txt
PyQt5==5.15.9
numpy==1.21.5
matplotlib==3.6.2
scipy==1.8.0
pylibtiff==0.5.1

